+++ tempodb/compaction_block_selector.go
@@ -40,12 +40,12 @@ type timeWindowBlockEntry struct {
 	meta  *backend.BlockMeta
 	group string // Blocks in the same group will be compacted together. Sort order also determines group priority.
 	order string // Individual block priority within the group.
-	hash  string // Hash string used for sharding ownership, preserves backwards compatibility
+	hash  string // hash string used for sharding ownership, preserves backwards compatibility
 }
 
 var _ (CompactionBlockSelector) = (*timeWindowBlockSelector)(nil)
 
-func newTimeWindowBlockSelector(blocklist []*backend.BlockMeta, maxCompactionRange time.Duration, maxCompactionObjects int, maxBlockBytes uint64, minInputBlocks int, maxInputBlocks int) CompactionBlockSelector {
+func newTimeWindowBlockSelector(blocklist []*backend.BlockMeta, maxCompactionRange time.Duration, maxCompactionObjects int, maxBlockBytes uint64, minInputBlocks, maxInputBlocks int) CompactionBlockSelector {
 	twbs := &timeWindowBlockSelector{
 		MinInputBlocks:       minInputBlocks,
 		MaxInputBlocks:       maxInputBlocks,
@@ -82,7 +82,8 @@ func newTimeWindowBlockSelector(blocklist []*backend.BlockMeta, maxCompactionRan
 
 			// Within group choose smallest blocks first.
 			// update after parquet: we want to make sure blocks of the same version end up together
-			entry.order = fmt.Sprintf("%016X-%v", entry.meta.TotalObjects, entry.meta.Version)
+			// update afert vParquet3: we want to make sure blocks of the same dedicated columns end up together
+			entry.order = fmt.Sprintf("%016X-%v-%016X", entry.meta.TotalObjects, entry.meta.Version, entry.meta.DedicatedColumnsHash())
 
 			entry.hash = fmt.Sprintf("%v-%v-%v", b.TenantID, b.CompactionLevel, w)
 		} else {
@@ -92,7 +93,8 @@ func newTimeWindowBlockSelector(blocklist []*backend.BlockMeta, maxCompactionRan
 
 			// Within group chose lowest compaction lvl and smallest blocks first.
 			// update after parquet: we want to make sure blocks of the same version end up together
-			entry.order = fmt.Sprintf("%v-%016X-%v", b.CompactionLevel, entry.meta.TotalObjects, entry.meta.Version)
+			// update afert vParquet3: we want to make sure blocks of the same dedicated columns end up together
+			entry.order = fmt.Sprintf("%v-%016X-%v-%016X", b.CompactionLevel, entry.meta.TotalObjects, entry.meta.Version, entry.meta.DedicatedColumnsHash())
 
 			entry.hash = fmt.Sprintf("%v-%v", b.TenantID, w)
 		}
@@ -127,6 +129,7 @@ func (twbs *timeWindowBlockSelector) BlocksToCompact() ([]*backend.BlockMeta, st
 				if twbs.entries[i].group == twbs.entries[j].group &&
 					twbs.entries[i].meta.DataEncoding == twbs.entries[j].meta.DataEncoding &&
 					twbs.entries[i].meta.Version == twbs.entries[j].meta.Version && // update after parquet: only compact blocks of the same version
+					twbs.entries[i].meta.DedicatedColumnsHash() == twbs.entries[j].meta.DedicatedColumnsHash() && // update after vParquet3: only compact blocks of the same dedicated columns
 					len(stripe) <= twbs.MaxInputBlocks &&
 					totalObjects(stripe) <= twbs.MaxCompactionObjects &&
 					totalSize(stripe) <= twbs.MaxBlockBytes {